requirements.txt
blinker==1.9.0
Brotli==1.1.0
cffi==1.17.1
chardet==5.2.0
click==8.1.8
colorama==0.4.6
cssselect2==0.8.0
dnspython==2.7.0
Flask==3.1.0
Flask-PyMongo==3.0.1
fonttools==4.56.0
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
packaging==24.2
pillow==11.1.0
pycparser==2.22
pydyf==0.11.0
PyJWT==2.10.1
pymongo==4.11.1
pyphen==0.17.2
pytz==2025.1
reportlab==4.3.1
tinycss2==1.4.0
tinyhtml5==2.0.0
webencodings==0.5.1
Werkzeug==3.1.3
zopfli==0.2.3.post1
